Separated PostArgs from FetchArgs
